+++ apps/dsp-app/src/app/project/project.component.ts
@@ -106,7 +106,8 @@ export class ProjectComponent implements OnInit {
      */
     @HostListener('window:keyup', ['$event'])
     keyEvent(event: KeyboardEvent) {
-        if (event.key === '[') {
+        const element = event.target as HTMLElement;
+        if (event.key === '[' && !element.matches('input, textarea')) {
             this.toggleSidenav();
         }
     }